Switch to Typescript
Close #5
