+++ package.json
@@ -13,9 +13,9 @@
     "build": "gatsby build",
     "dev": "gatsby develop",
     "local:prod": "gatsby build && gatsby serve",
-    "format": "prettier --write 'src/**/*.js'",
+    "format": "prettier --write 'src/**/*+(ts|tsx)?'",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint ./src",
+    "lint": "eslint ./src --ext .ts --ext .tsx",
     "deploy": "gatsby build && gh-pages -d public -b master -r git@github.com:ValentinH/valentinh.github.io.git"
   },
   "dependencies": {
@@ -26,6 +26,8 @@
     "@fortawesome/free-solid-svg-icons": "^5.1.0-11",
     "@fortawesome/react-fontawesome": "^0.1.0-11",
     "@sentry/browser": "^4.5.3",
+    "@types/react-helmet": "^5.0.8",
+    "@types/react-image-gallery": "^0.8.3",
     "animated-scroll-to": "^1.2.2",
     "axios": "^0.18.0",
     "gatsby": "^2.0.98",
@@ -38,6 +40,7 @@
     "gatsby-plugin-prefetch-google-fonts": "^1.1.1",
     "gatsby-plugin-react-helmet": "^3.0.5",
     "gatsby-plugin-sentry": "^1.0.1",
+    "gatsby-plugin-typescript": "^2.0.3",
     "google-fonts-plugin": "2.0.2",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
@@ -47,14 +50,16 @@
     "react-intersection-observer": "^6.4.2"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.6",
-    "eslint": "^5.3.0",
-    "eslint-config-prettier": "^3.0.1",
+    "@typescript-eslint/eslint-plugin": "^1.1.0",
+    "@typescript-eslint/parser": "^1.1.0",
+    "eslint": "^5.12.1",
+    "eslint-config-prettier": "^4.0.0",
     "eslint-plugin-emotion": "^10.0.6",
-    "eslint-plugin-prettier": "^2.6.2",
-    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-react": "^7.12.4",
     "gh-pages": "^2.0.1",
-    "prettier": "^1.13.7"
+    "prettier": "^1.16.1",
+    "typescript": "^3.2.4"
   },
   "resolutions": {
     "*/google-fonts-plugin": "2.0.2"
