+++ src/components/Content/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
-
 import { Container } from './styles'
 
-const Content = ({ children }) => <Container>{children}</Container>
+const Content: React.SFC = ({ children }) => <Container>{children}</Container>
 
 export default Content
