+++ src/components/Content/styles.ts
