+++ src/components/Education/styles.ts
