+++ src/components/Footer/styles.ts
