+++ src/components/ImagesGallery/styles.ts
