+++ src/components/Intro/styles.ts
