+++ src/components/Projects/styles.ts
