+++ src/declaration.d.ts
@@ -1,0 +1,14 @@
+declare module '*.png' {
+  const content: any
+  export default content
+}
+
+declare module '*.jpg' {
+  const content: any
+  export default content
+}
+
+declare module '*.webp' {
+  const content: any
+  export default content
+}
